Move Output Widget to New File; Implement Block System
Since the Output widget's systems are large enough to move into a separate file, this has been done.

In addition, a new text block system has been added to deal with dynamic text. Instead of needing a 'dynamic' flag and any number of other possible flags in the 'display' method, the 'display' method now returns a Block object that represents a block of text. This Block object allows the user to handle these sort of flags without needing to pass these flags to one method.

Signed-off-by: Tate Mioton <tatemioton@outlook.com>